backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
google-generativeai
groq
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
httpx
yt-dlp
celery
redis
